refactor: move TxOutProof into fedimint-common
