+++ gateway/cli/src/main.rs
@@ -2,9 +2,9 @@ use std::process::exit;
 
 use bitcoin::{Address, Amount, Transaction};
 use clap::{Parser, Subcommand};
-use fedimint_client_legacy::modules::wallet::txoproof::TxOutProof;
 use fedimint_client_legacy::utils::from_hex;
 use fedimint_core::config::FederationId;
+use fedimint_core::txoproof::TxOutProof;
 use fedimint_logging::TracingSetup;
 use ln_gateway::rpc::rpc_client::RpcClient;
 use ln_gateway::rpc::{
