+++ gateway/ln-gateway/src/actor.rs
@@ -7,9 +7,9 @@ use bitcoin_hashes::{sha256, Hash};
 use fedimint_client_legacy::mint::backup::Metadata;
 use fedimint_client_legacy::modules::ln::contracts::{ContractId, Preimage};
 use fedimint_client_legacy::modules::ln::route_hints::RouteHint;
-use fedimint_client_legacy::modules::wallet::txoproof::TxOutProof;
 use fedimint_client_legacy::{GatewayClient, PaymentParameters};
 use fedimint_core::task::{RwLock, TaskGroup};
+use fedimint_core::txoproof::TxOutProof;
 use fedimint_core::{Amount, OutPoint, TransactionId};
 use futures::stream::{BoxStream, StreamExt};
 use rand::{CryptoRng, RngCore};
