+++ gateway/ln-gateway/src/rpc/mod.rs
@@ -8,8 +8,8 @@ use anyhow::{anyhow, Error};
 use bitcoin::{Address, Transaction, XOnlyPublicKey};
 use bitcoin_hashes::hex::{FromHex, ToHex};
 use fedimint_client_legacy::ln::PayInvoicePayload;
-use fedimint_client_legacy::modules::wallet::txoproof::TxOutProof;
 use fedimint_core::config::FederationId;
+use fedimint_core::txoproof::TxOutProof;
 use fedimint_core::{Amount, TransactionId};
 use futures::Future;
 use serde::{Deserialize, Deserializer, Serialize, Serializer};
