+++ integrationtests/tests/fixtures/legacy.rs
@@ -21,11 +21,12 @@ use fedimint_core::db::Database;
 use fedimint_core::epoch::SignedEpochOutcome;
 use fedimint_core::module::registry::ModuleDecoderRegistry;
 use fedimint_core::task::TaskGroup;
+use fedimint_core::txoproof::TxOutProof;
 use fedimint_core::{Amount, OutPoint, PeerId, TieredMulti, TransactionId};
 use fedimint_ln_client::contracts::{ContractId, Preimage};
 use fedimint_ln_client::{ContractAccount, LightningGateway};
 use fedimint_mint_client::BlindNonce;
-use fedimint_wallet_client::txoproof::{PegInProof, TxOutProof};
+use fedimint_wallet_client::txoproof::PegInProof;
 use fedimint_wallet_client::{PegOut, PegOutFees, Rbf};
 use futures::executor::block_on;
 use lightning_invoice::Invoice;
