+++ integrationtests/tests/fixtures/real.rs
@@ -13,9 +13,9 @@ use cln_rpc::{ClnRpc, Request, Response};
 use fedimint_bitcoind::DynBitcoindRpc;
 use fedimint_core::encoding::Decodable;
 use fedimint_core::module::registry::ModuleDecoderRegistry;
+use fedimint_core::txoproof::TxOutProof;
 use fedimint_core::Amount;
 use fedimint_testing::btc::BitcoinTest;
-use fedimint_wallet_server::common::txoproof::TxOutProof;
 use lazy_static::lazy_static;
 use lightning_invoice::Invoice;
 use tokio::sync::Mutex;
