+++ integrationtests/tests/fixtures/user.rs
@@ -12,11 +12,12 @@ use fedimint_core::config::ClientConfig;
 use fedimint_core::core::KeyPair;
 use fedimint_core::epoch::SignedEpochOutcome;
 use fedimint_core::task::TaskGroup;
+use fedimint_core::txoproof::TxOutProof;
 use fedimint_core::{Amount, OutPoint, PeerId, TieredMulti, TransactionId};
 use fedimint_ln_client::contracts::{ContractId, Preimage};
 use fedimint_ln_client::{ContractAccount, LightningGateway};
 use fedimint_mint_client::BlindNonce;
-use fedimint_wallet_client::txoproof::{PegInProof, TxOutProof};
+use fedimint_wallet_client::txoproof::PegInProof;
 use fedimint_wallet_client::{PegOut, PegOutFees, Rbf};
 use futures::executor::block_on;
 use itertools::Itertools;
