+++ modules/fedimint-wallet-common/src/txoproof.rs
@@ -1,17 +1,13 @@
-use std::borrow::Cow;
 use std::convert::Infallible;
 use std::hash::Hash;
-use std::io::Cursor;
 
-use bitcoin::hashes::hex::{FromHex, ToHex};
-use bitcoin::util::merkleblock::PartialMerkleTree;
-use bitcoin::{BlockHash, BlockHeader, OutPoint, Transaction, Txid};
+use bitcoin::{BlockHash, OutPoint, Transaction};
 use fedimint_core::encoding::{Decodable, DecodeError, Encodable};
 use fedimint_core::module::registry::ModuleDecoderRegistry;
+use fedimint_core::txoproof::TxOutProof;
 use miniscript::{Descriptor, TranslatePk};
 use secp256k1::{Secp256k1, Signing, Verification};
-use serde::de::Error;
-use serde::{Deserialize, Deserializer, Serialize, Serializer};
+use serde::{Deserialize, Serialize};
 use thiserror::Error;
 use validator::{Validate, ValidationError};
 
@@ -31,69 +27,6 @@ pub struct PegInProof {
     // Check that the idx is in range
     output_idx: u32,
     tweak_contract_key: secp256k1::XOnlyPublicKey,
-}
-
-#[derive(Clone, Debug)]
-pub struct TxOutProof {
-    pub block_header: BlockHeader,
-    pub merkle_proof: PartialMerkleTree,
-}
-
-impl TxOutProof {
-    pub fn block(&self) -> BlockHash {
-        self.block_header.block_hash()
-    }
-
-    pub fn contains_tx(&self, tx_id: Txid) -> bool {
-        let mut transactions = Vec::new();
-        let mut indices = Vec::new();
-        let root = self
-            .merkle_proof
-            .extract_matches(&mut transactions, &mut indices)
-            .expect("Checked at construction time");
-
-        debug_assert_eq!(root, self.block_header.merkle_root);
-
-        transactions.contains(&tx_id)
-    }
-}
-
-impl Decodable for TxOutProof {
-    fn consensus_decode<D: std::io::Read>(
-        d: &mut D,
-        modules: &ModuleDecoderRegistry,
-    ) -> Result<Self, DecodeError> {
-        let block_header = BlockHeader::consensus_decode(d, modules)?;
-        let merkle_proof = PartialMerkleTree::consensus_decode(d, modules)?;
-
-        let mut transactions = Vec::new();
-        let mut indices = Vec::new();
-        let root = merkle_proof
-            .extract_matches(&mut transactions, &mut indices)
-            .map_err(|_| DecodeError::from_str("Invalid partial merkle tree"))?;
-
-        if block_header.merkle_root != root {
-            Err(DecodeError::from_str(
-                "Partial merkle tree does not belong to block header",
-            ))
-        } else {
-            Ok(TxOutProof {
-                block_header,
-                merkle_proof,
-            })
-        }
-    }
-}
-
-impl Encodable for TxOutProof {
-    fn consensus_encode<W: std::io::Write>(&self, writer: &mut W) -> Result<usize, std::io::Error> {
-        let mut written = 0;
-
-        written += self.block_header.consensus_encode(writer)?;
-        written += self.merkle_proof.consensus_encode(writer)?;
-
-        Ok(written)
-    }
 }
 
 impl PegInProof {
@@ -211,45 +144,6 @@ impl Tweakable for Descriptor<CompressedPublicKey> {
     }
 }
 
-impl Serialize for TxOutProof {
-    fn serialize<S>(&self, serializer: S) -> Result<<S as Serializer>::Ok, <S as Serializer>::Error>
-    where
-        S: Serializer,
-    {
-        let mut bytes = Vec::new();
-        self.consensus_encode(&mut bytes).unwrap();
-
-        if serializer.is_human_readable() {
-            serializer.serialize_str(&bytes.to_hex())
-        } else {
-            serializer.serialize_bytes(&bytes)
-        }
-    }
-}
-
-impl<'de> Deserialize<'de> for TxOutProof {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        let empty_module_registry = ModuleDecoderRegistry::default();
-        if deserializer.is_human_readable() {
-            let hex_str: Cow<str> = Deserialize::deserialize(deserializer)?;
-            let bytes = Vec::from_hex(&hex_str).map_err(D::Error::custom)?;
-            Ok(
-                TxOutProof::consensus_decode(&mut Cursor::new(bytes), &empty_module_registry)
-                    .map_err(D::Error::custom)?,
-            )
-        } else {
-            let bytes: &[u8] = Deserialize::deserialize(deserializer)?;
-            Ok(
-                TxOutProof::consensus_decode(&mut Cursor::new(bytes), &empty_module_registry)
-                    .map_err(D::Error::custom)?,
-            )
-        }
-    }
-}
-
 fn validate_peg_in_proof(proof: &PegInProof) -> Result<(), ValidationError> {
     if !proof.txout_proof.contains_tx(proof.transaction.txid()) {
         return Err(ValidationError::new(
@@ -276,23 +170,6 @@ fn validate_peg_in_proof(proof: &PegInProof) -> Result<(), ValidationError> {
 
     Ok(())
 }
-
-// TODO: upstream
-impl Hash for TxOutProof {
-    fn hash<H: std::hash::Hasher>(&self, state: &mut H) {
-        let mut bytes = Vec::new();
-        self.consensus_encode(&mut bytes).unwrap();
-        state.write(&bytes);
-    }
-}
-
-impl PartialEq for TxOutProof {
-    fn eq(&self, other: &TxOutProof) -> bool {
-        self.block_header == other.block_header && self.merkle_proof == other.merkle_proof
-    }
-}
-
-impl Eq for TxOutProof {}
 
 impl Decodable for PegInProof {
     fn consensus_decode<D: std::io::Read>(
@@ -330,8 +207,7 @@ mod tests {
     use bitcoin::hashes::hex::FromHex;
     use fedimint_core::encoding::Decodable;
     use fedimint_core::module::registry::ModuleDecoderRegistry;
-
-    use super::TxOutProof;
+    use fedimint_core::txoproof::TxOutProof;
 
     #[test_log::test]
     fn test_txoutproof_happy_path() {
